+++ src/hooks/useWatchGetter.ts
@@ -1,15 +1,19 @@
-import { useState, useEffect } from "react";
+import { useState, useEffect, useRef } from "react";
 
-function WatchGetterAnimation<T>(getter: () => T): T {
+function WatchGetterAnimation<T>(getter: () => T, callback: (v: T) => void, updater?: boolean):
+  typeof updater extends true ? [T, () => void] : T {
   const [value, setValue] = useState(getter());
+  const valueRef = useRef(value);
 
   useEffect(() => {
     let animationFrameId: number;
 
     const loop = () => {
       const newValue = getter();
-      if (newValue !== value) {
+      if (newValue !== valueRef.current) {
+        valueRef.current = newValue;
         setValue(newValue);
+        callback?.(newValue);
       }
       animationFrameId = requestAnimationFrame(loop);
     };
@@ -19,54 +23,26 @@ function WatchGetterAnimation<T>(getter: () => T): T {
     return () => {
       cancelAnimationFrame(animationFrameId);
     };
-  }, [getter, value]);
+  }, [getter]);
 
-  return value;
-}
-
-function WatchGetterSetter<T>(
-  getter: () => T,
-  ...setters: ((value: any) => void)[]
-): T {
-  const [value, setValue] = useState(getter());
-
-  useEffect(() => {
-    const _setters = setters.map((setter) => setter);
-    _setters.forEach((setter) => {
-      let originSetter = setter;
-      // @ts-ignore
-      originSetter = (...args) => {
-        // @ts-ignore
-
-        setValue(getter());
-        originSetter(...args);
-      };
-    });
-
-    return () => {
-      _setters.forEach((setter) => {
-        let originSetter = setter;
-        // @ts-ignore
-        originSetter = (...args) => {
-          // @ts-ignore
-          originSetter(...args);
-        };
-      });
-    };
-  }, [getter, setters]);
-
-  return value;
+  const update = () => {
+    const newValue = getter();
+    if (newValue !== valueRef.current) {
+      valueRef.current = newValue;
+      setValue(newValue);
+      callback?.(newValue);
+    }
+  }
+  // @ts-ignore
+  return updater ? [value, update] : value;
 }
 
 function useWatchGetter<T>(
   getter: () => T,
-  ...setters: ((value: any) => void)[]
-): T {
-  if (setters.length === 0) {
-    return WatchGetterAnimation(getter);
-  } else {
-    return WatchGetterSetter(getter, ...setters);
-  }
+  callback: (v: T) => void = () => { },
+  updater: boolean = false
+): typeof updater extends true ? [T, () => void] : T {
+  return WatchGetterAnimation(getter, callback, updater);
 }
 
 export default useWatchGetter;