+++ src/views/docs/views/examples/UseWatchGetter/example.1.tsx
@@ -1,0 +1,42 @@
+import useWatchGetter from "@hooks/useWatchGetter";
+import { Button } from "@mui/material";
+import { useState } from "react";
+
+const View = () => {
+  const [obj, setObj] = useState({ a: 1, b: 2 });
+  const increment = () => setObj((prev) => ({ ...prev, a: prev.a + 1 }));
+  const decrement = () => setObj((prev) => ({ ...prev, a: prev.a - 1 }));
+  const value = useWatchGetter(() => obj.a);
+  return (
+    <>
+      <Button onClick={increment}>Increment</Button>
+      <Button onClick={decrement}>Decrement</Button>
+      <pre>{JSON.stringify(value, null, 2)}</pre>
+    </>
+  );
+};
+
+const code = `
+import { useWatchGetter } from "@evanpatchouli/react-hooks-kit";
+import { Button } from "@mui/material";
+import { useState } from "react";
+
+const View = () => {
+  const [obj, setObj] = useState({ a: 1, b: 2 });
+  const increment = () => setObj((prev) => ({ ...prev, a: prev.a + 1 }));
+  const decrement = () => setObj((prev) => ({ ...prev, a: prev.a - 1 }));
+  const value = useWatchGetter(() => obj.a);
+  return (
+    <>
+      <Button onClick={increment}>Increment</Button>
+      <Button onClick={decrement}>Decrement</Button>
+      <pre>{JSON.stringify(value, null, 2)}</pre>
+    </>
+  );
+};
+`;
+
+export default {
+  code,
+  View,
+};